Extend pipeline at routing layer.
