+++ src/Illuminate/Routing/Pipeline.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace Illuminate\Routing;
+
+use Exception;
+use Illuminate\Http\Request;
+use Illuminate\Contracts\Debug\ExceptionHandler;
+use Illuminate\Pipeline\Pipeline as BasePipeline;
+use Symfony\Component\Debug\Exception\FatalThrowableError;
+
+class Pipeline extends BasePipeline
+{
+    /**
+     * Get a Closure that represents a slice of the application onion.
+     *
+     * @return \Closure
+     */
+    protected function getSlice()
+    {
+        return function ($stack, $pipe) {
+            return function ($passable) use ($stack, $pipe) {
+                try {
+                    return call_user_func(parent::getSlice()($stack, $pipe), $passable);
+                } catch (Exception $e) {
+                    return $this->handleException($passable, $e);
+                } catch (Throwable $e) {
+                    return $this->handleException($passable, new FatalThrowableError($e));
+                }
+            };
+        };
+    }
+
+    /**
+     * Handle the given exception.
+     *
+     * @param  \Exception  $e
+     * @return mixed
+     */
+    protected function handleException($passable, Exception $e)
+    {
+        if (! $this->container->bound(ExceptionHandler::class) ||
+            ! $passable instanceof Request) {
+            throw $e;
+        }
+
+        $handler = $this->container->make(ExceptionHandler::class);
+
+        $handler->report($e);
+
+        return $handler->render($passable, $e);
+    }
+}
