+++ src/Illuminate/Routing/composer.json
@@ -21,6 +21,7 @@
         "illuminate/pipeline": "5.2.*",
         "illuminate/session": "5.2.*",
         "illuminate/support": "5.2.*",
+        "symfony/debug": "2.8.*|3.0.*",
         "symfony/http-foundation": "2.8.*|3.0.*",
         "symfony/http-kernel": "2.8.*|3.0.*",
         "symfony/routing": "2.8.*|3.0.*"
